Null fix in export #3919

+++ src/main/java/edu/harvard/iq/dataverse/export/ExportService.java
@@ -276,9 +276,10 @@ public class ExportService {
                     System.out.println("Saving path as aux for temp file in: " + Paths.get(tempFile.getAbsolutePath()));
                     storageIO.savePathAsAux(Paths.get(tempFile.getAbsolutePath()), "export_" + format + ".cached");
                 }
-                
-                boolean tempFileDeleted = tempFile.delete();
-                logger.fine("tempFileDeleted: " + tempFileDeleted);
+                if (tempFile != null) {
+                    boolean tempFileDeleted = tempFile.delete();
+                    logger.fine("tempFileDeleted: " + tempFileDeleted);
+                }
 
             } catch (IOException ioex) {
                 throw new ExportException("IO Exception thrown exporting as " + "export_" + format + ".cached");
